VEA-34 Renamed code.exceptions package.
Renamed package com.danieldigiovanni.email.code.exceptions to
com.danieldigiovanni.email.code.exception.
